Update 8
Created dice program that returns a random number between 1 and 6 when called.   Updated family generator program: function that converts numbers to percentages of total families.  Added function that assigns numbers of boys or girls to separate arrays in correct places.  Added condition that adds number of boys and girls and places them in separate arrays in their correct positions based on # children per family 0 - 6.  Added logs that present a readable format to the data based on frequency of children in each group  0 - 6 out of total family population.  Updated mutation sequence program to check if an individual base already exists during mutation and if so generate a different random base until the bases are different.