Update 10
Created allele people program that creates a random population of people with different alleles and functions to grab random individual alleles from specific individuals within the Population.

Added notes on mutation to the mutation-seq program.

Created a spatial mating simulation.  Created a function that creates a 2d grid of arrays and populates them with random genotypes based on specific allele frequencies.  Created a run generation function that creates the first generation by looping through the main grid, picking a mating partner, reproducing, and mapping the new offspring to the same location as parent on the temp grid.  Created a loop that goes through the main grid and overwrites it with the new temp grid of offspring.  Created a loop that runs 100 generations.