Update 15
Added notes on natural selection.  Updated the coevolution program to properly push frequency arrays into the data array.  Added small random variation to host and parasite frequency.

Created epidemics program that creates a grid of 100 individuals and assigns an "S" for susceptable tag to each.  The program assigns one random of the individuals an "I" for infected.  It then draws a the grid and visualizes the different status types of individuals with different colors.